Implement loading IntegrationTestSummary from json
Also, remove the parsers directory, move the integration test parser
back to github-ci, and separate out the JUnitXMLParser and
IntegrationTestSummary classes
